backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.3
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
sniffio==1.3.1
spotipy==2.24.0
starlette==0.41.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
